+++ Mods ChoGGi/Fix Remove Blue Yellow Marks And Ghosts/Code/Script.lua
@@ -1,5 +1,7 @@
 -- See LICENSE for terms
 
+local DoneObject = DoneObject
+
 local mod_EnableMod
 
 local function ModOptions(id)
@@ -14,6 +16,12 @@ end
 OnMsg.ModsReloaded = ModOptions
 -- Fired when Mod Options>Apply button is clicked
 OnMsg.ApplyModOptions = ModOptions
+
+local rover_ents = {
+	RoverBlueSunHarvester = true,
+	RoverIndiaConstructor = true,
+	RoverTransport = true,
+}
 
 function OnMsg.LoadGame()
 	if not mod_EnableMod then
@@ -30,18 +38,17 @@ function OnMsg.LoadGame()
 			return true
 		end
 	end)
+
 	for i = #objs, 1, -1 do
-		objs[i]:delete()
+		DoneObject(objs[i])
 	end
 
 	-- remove the rover outlines added from https://forum.paradoxplaza.com/forum/index.php?threads/surviving-mars-persistent-transport-route-blueprint-on-map.1121333/
-	objs = MapGet(true, "WireFramedPrettification", function(o)
-		if o:GetEntity() == "RoverTransport" then
-			return true
-		end
+	objs = MapGet(true, "WireFramedPrettification", function(rover)
+		return rover_ents[rover.entity or ""]
 	end)
 	for i = #objs, 1, -1 do
-		objs[i]:delete()
+		DoneObject(objs[i])
 	end
 
 	ResumePassEdits("ChoGGi.RemoveBlueYellowGridMarks.LoadGame")