+++ Mods ChoGGi/Fix Remove Blue Yellow Marks And Ghosts/metadata.lua
@@ -3,9 +3,9 @@ return PlaceObj("ModDef", {
 	"id", "ChoGGi_FixRemoveBlueYellowGridMarks",
 	"steam_id", "1553086208",
 	"pops_any_uuid", "7486c4ab-af06-42d5-b30d-7815ac10998b",
-	"version", 3,
+	"version", 4,
 	"version_major", 0,
-	"version_minor", 3,
+	"version_minor", 4,
 	"image", "Preview.jpg",
 	"author", "ChoGGi",
 	"lua_revision", 1007000, -- Picard
@@ -14,7 +14,8 @@ return PlaceObj("ModDef", {
 	},
 	"has_options", true,
 	"TagOther", true,
-	"description", [[This will remove any blue or yellow grid marks from around objects when you load a game.
+	"description", [[
+This will remove any blue or yellow grid marks from around objects when you load a game.
 This also takes care of stuck "ghost" rovers from transport tasks.
 Includes mod option to disable.